halocline-surrogate/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "halocline-surrogate"
version = "0.1.0"
description = "Surrogate modeling package for the Halocline Python physics oracle."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "halocline-physics",
  "h5py>=3.10",
  "numpy>=1.26",
  "PyYAML>=6",
  "torch>=2.2",
]

[project.optional-dependencies]
report = ["matplotlib>=3.8"]
test = ["pytest>=8"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "../halocline-physics/src"]
